end of section 13

+++ App/controllers/ListingController.php
@@ -4,6 +4,9 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
+use Framework\Middleware\Authorize;
 
 class ListingController
 {
@@ -23,7 +26,7 @@ class ListingController
     public function index()
 
     {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
 
         loadView('listings/index', [
@@ -81,7 +84,7 @@ class ListingController
         // to match the keys
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         // to sanitize
         $newListingData = array_map('sanitize', $newListingData);
@@ -128,7 +131,7 @@ class ListingController
             $this->db->query($query, $newListingData);
 
             //Set Flash message
-            $_SESSION['succes_message'] = 'Listing created succesfully';
+            Session::setFlashMessage('success_message', 'Listing created successfully');
 
             redirect('/listings');
         }
@@ -149,15 +152,22 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM Listings WHERE id = :id', $params)->fetch();
 
+        //  Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/listings/' . $listing->id);
+        }
+
         $this->db->query('DELETE FROM Listings WHERE id= :id', $params);
 
         //Set Flash message
-        $_SESSION['succes_message'] = 'Listing deleted succesfully';
+        Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
         redirect('/listings');
     }
@@ -182,6 +192,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing->id);
         }
 
         loadView('listings/edit', [
@@ -209,6 +225,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing->id);
         }
 
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits',];
@@ -251,7 +273,9 @@ class ListingController
             $updateValues['id'] = $id;
             $this->db->query($updateQuery, $updateValues);
 
-            $_SESSION['succes_message'] = 'Listing Updated';
+            //Set Flash message
+            Session::setFlashMessage('success_message', 'Listing updated successfully');
+
 
             redirect('/listings/' . $id);
         }
